Name:Aryaman Rawat Date:18/11/2022
Added new method pop() for data type float in (NumStack.java).This method
was implemented by creating a new test case to check for empty stack
exception and if popping from a empty stack throws the error.

Major refactoring was done to (NumStack.java) the checkstyles are upto
the requirment and all the required java doc comments have been made to
the codes.

All main methods implemented on (NumStack.java).More testing to be done
before pushing to main branch.